refactor: moving the the resoure into it own file
