+++ backend/zane_api/views/__init__.py
@@ -7,3 +7,4 @@ from .logs import *
 from .projects import *
 from .proxy import *
 from .status import *
+from .search import *
