+++ backend/zane_api/views/projects.py
@@ -197,56 +197,6 @@ class ArchivedProjectsListAPIView(ListAPIView):
             return Response(EMPTY_PAGINATED_RESPONSE)
 
 
-class ResouceSearchAPIView(APIView):
-    @extend_schema(
-        operation_id="searchResources",
-        summary="search for resources (project, service ...)",
-        parameters=[ResourceSearchParamSerializer],
-        responses={
-            200: PolymorphicProxySerializer(
-                component_name="ResourceResponse",
-                serializers=[
-                    ServiceSearchSerializer,
-                    ProjectSearchSerializer,
-                ],
-                resource_type_field_name="type",
-                many=True,
-            )
-        },
-    )
-    def get(self, request: Request) -> Response:
-        query = request.query_params.get("query", "").strip()
-        projects = Project.objects.filter(slug__istartswith=query)[:5]
-        projects_object = [
-            {"id": project.id, "slug": project.slug, "created_at": project.created_at}
-            for project in projects
-        ]
-
-        services: QuerySet[DockerRegistryService] = (
-            DockerRegistryService.objects.filter(
-                slug__istartswith=query
-            ).select_related("project")[:5]
-        )
-
-        services_object = [
-            {
-                "id": service.id,
-                "slug": service.slug,
-                "created_at": service.created_at,
-                "project_slug": service.project.slug,
-            }
-            for service in services
-        ]
-
-        return Response(
-            [
-                *ProjectSearchSerializer(projects_object, many=True).data,
-                *ServiceSearchSerializer(services_object, many=True).data,
-            ],
-            status=status.HTTP_200_OK,
-        )
-
-
 class ProjectDetailsView(APIView):
     serializer_class = ProjectSerializer
